basic warnings cleanup (not touching object's memory layout). A
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/353

basic build now compiles w/o a single warning on Linux. One with --enable-strict
is a different matter...


git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@1124 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/common/convert.cpp
@@ -61,10 +61,10 @@ UnicodeConverterCPP&   UnicodeConverterCPP::operator=(const UnicodeConverterCPP&
          *Increments the assigner converter's ref count
          */
       Mutex updateReferenceCounters;
-      if (myUnicodeConverter->sharedData->referenceCounter != 0 && myUnicodeConverter->sharedData->referenceCounter != ~0) {
+      if (myUnicodeConverter->sharedData->referenceCounter != 0 && myUnicodeConverter->sharedData->referenceCounter != (uint32_t) ~0) {
         myUnicodeConverter->sharedData->referenceCounter--;
       }
-      if (that.myUnicodeConverter->sharedData->referenceCounter != ~0) {
+      if (that.myUnicodeConverter->sharedData->referenceCounter != (uint32_t) ~0) {
         that.myUnicodeConverter->sharedData->referenceCounter++;
       }
     }
@@ -102,7 +102,7 @@ UnicodeConverterCPP::UnicodeConverterCPP(const UnicodeConverterCPP&  that)
     myUnicodeConverter = new UConverter;
     {
       Mutex updateReferenceCounter;
-      if (that.myUnicodeConverter->sharedData->referenceCounter != ~0) {
+      if (that.myUnicodeConverter->sharedData->referenceCounter != (uint32_t) ~0) {
         that.myUnicodeConverter->sharedData->referenceCounter++;
       }
     }
